+++ HISTORY.md
@@ -3,6 +3,7 @@
 ## 0.2.0
 
 1. WARNING: All the public API fns have been renamed.
+2. Added unit tests.
 
 ## 0.1.0